requirements.txt
Flask==3.1.0
Pillow==11.1.0
numpy==2.2.4
Werkzeug==3.1.3
